requirements.txt
Django==4.1.7
Pillow==9.5.0
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0